duplicados en el carrito
Eliminación de productos duplicados en el carrito de compras

+++ app/Http/Controllers/CartDetailController.php
@@ -9,20 +9,26 @@ class CartDetailController extends Controller
 {
 	public function store(Request $request)
 	{
-
-		$cartDetail = new CartDetail();
-		$cartDetail->cart_id = auth()->user()->cart->id;
-		$cartDetail->product_id=$request->product_id;
-		$cartDetail->quantity=$request->quantity;
-		$cartDetail->save();
-
+		$cart = auth()->user()->cart;
+		$detail=$cart->details()->where('product_id', $request->product_id)->first();
+		if($detail){
+			$detail->quantity += $request->quantity;
+			$detail->save();
+		}else{
+			$cartDetail = new CartDetail();
+			$cartDetail->cart_id =auth()->user()->cart->id;
+			$cartDetail->product_id = $request->product_id;
+			$cartDetail->quantity = $request->quantity;
+			$cartDetail->save();
+		}
 		return back();
 	}
 	public function destroy(Request $request)
 	{
 
 		$cartDetail = CartDetail::find($request->cart_detail_id);
-		$cartDetail->delete();
+		if ($cartDetail->cart_id == auth()->user()->cart->id)
+			$cartDetail->delete();
 
 		return back();
 	}
